[20.3%] Seperate API.ex into command / query concerns
